refactor [DEV-27]: Update imports for the angular ui module
