+++ libs/angular/ui/src/modules/media/index.ts
@@ -28,3 +28,4 @@ export {
 	mediaBetweenGuard,
 	mediaMinGuard,
 } from './guards';
+export { MEDIA_DEVICE, MEDIA_CONFIG, MEDIA_ELEMENT } from './tokens';
